Move README files to a dedicated docs directory
The new dedicated docs directory has been introduced after a discussion
on GitHub[1].

Main issue it is addressing is the reduction of too many README files in
the project root directory. The new directory is dedicated for notes and
quick documentation files that either can't be put in the manual or wiki
pages or that relate to the php-src repository specifically and need to
live together with the source code.

Also the `docs` directory is by GitHub used for some repository
configuration files such as pull request templates, and contributing
documentation helper files that are integrated in the interface.

[1]: https://github.com/php/php-src/pull/3988